iter24 Выделил server env в отдельный файл
