+++ internal/server/docs.go
@@ -1,0 +1,3 @@
+// Package server реализует серверную часть приложения(за исключением Storage).
+// Содержит прежде всего хэндлеры и миддлвары, а также функциональность необходимую для работы сервера.
+package server
